requirements.txt
pandas>=2.0.0
jinja2>=3.1.0
matplotlib>=3.7.0
weasyprint>=60.0
pdfkit>=1.0.0
openpyxl>=3.1.0

# API Integrations (install these to use --api mode)
google-analytics-data>=0.18.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
facebook-business>=19.0.0

# Web App
flask>=3.0.0
gunicorn>=22.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
werkzeug>=3.0.0
